Add UniqueBlockNode interface
